+++ meshapp/src/pages/login/index.tsx
@@ -2,7 +2,7 @@ import { Button, Divider, Grid, Skeleton, Typography } from "@mui/material";
 
 import { paths } from "../../routing/RoutePaths";
 import { Link as RouterLink, useNavigate } from "react-router-dom";
-import LoginWindow from "../../components/login-window";
+import LoginWindow from "../../components/LoginWindow";
 import { GridContainer, GridItem } from "../../components/resuables/Grids";
 
 const SignUp = () => {
